"Modulo de Servicios Terminado" Tambien se tocaron varios archivos del config y otros mas para el funcionamiento del mismo modulo (servicios)

+++ application/modules/Mipanel/controllers/Nosotros.php
@@ -48,7 +48,7 @@ public function deleteImg()
     
     // recorremos los ficheros a ver si existe en la carpeta
     foreach ($ficheros as $img) {
-      if ($img === $fileName) {
+      if ($img === $fileName && $img != '400x300.jpg') {
         unlink($deletefile);
         $data['success'] = TRUE;
         break;
